% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGS.R
\name{BGS}
\alias{BGS}
\title{Bayesian Gibbs Sampling method}
\usage{
BGS(
  yt,
  Kt,
  R0 = NULL,
  mu.ini = NULL,
  ga.ini = 0,
  vE.ini = 1,
  vA.ini = 0.5,
  nu_star = 5,
  nIters = 5000,
  tolParInv = 10^-4,
  verbose = TRUE
)
}
\arguments{
\item{yt}{A numeric vector of phenotypic values corresponding to individuals of allocation of a training set.}

\item{Kt}{The corresponding Kinship matrix of the training set.}

\item{R0}{The known spatial correlation matrix for a specific field. Consider it independent for units if NULL.}

\item{mu.ini}{Initial values for the constant term.}

\item{ga.ini}{Initial values for the additive genotypic values.}

\item{vE.ini}{Initial values for the error variance.}

\item{nu_star}{The degree of freedom of the variance components for the prior distribution of a scaled inverse Chi square.}

\item{nIters}{Number of iterations.}

\item{tolParInv}{A tolerance parameter of matrix inversion.}

\item{verbose}{Show details in console if TRUE and vice versa.}

\item{va.ini}{Initial values for the additive variance.}
}
\value{
This function will return a list containing estimates of ga, mu, Va, and VE for each iteration.
}
\description{
A Bayesian Gibbs Sampling method to obtain GEBVs for GBLUP model.
BGS for additive spatial GBLUP model:
yt ~ mu\emph{1 + gt + e0, where
yt is phenotypic values for a given training population;
mu is fixed parameter of population mean;
1 denotes column vector with length nt;
gt is genotypic values of corresponding individuals for training set;                       # gt ~ MVN(0, sigma_g^2}Kt)
e0 is random error for a specific field with known spatially correlated experimental units. # e0 ~ MVN(0, sigma_e^2*R0)
}
