% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDmean.v2.R
\name{CDmean.v2}
\alias{CDmean.v2}
\title{CDmean(v2) criterion}
\usage{
CDmean.v2(Kb, train, test, R = NULL, lambda = NULL, tolParInv = 10^-4)
}
\arguments{
\item{Kb}{Kinship matrix of breeding population.}

\item{train}{An integer vector indicating corresponding allocation of a training population.}

\item{test}{An integer vector of targeted population.}

\item{R}{A nt*nt covariance matrix of random error for spatial correlation.}

\item{lambda}{A variance parameter representing the ratio of error variance to additive genetic variance for the computation of CDmean. Defaulted to 1.}

\item{tolParInv}{A tolerance parameter of matrix inversion.}
}
\description{
Compute the CDmean.v2 value for a given training set.
}
