% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEBV.R
\name{GEBV}
\alias{GEBV}
\title{GEBV prediction}
\usage{
GEBV(
  Kb,
  train,
  yt,
  R0 = NULL,
  nIters = 5000,
  tolParInv = 10^-4,
  method = "BGS"
)
}
\arguments{
\item{Kb}{A nb*nb Kinship matrix of breeding population.}

\item{train}{An integer vector indicating corresponding allocation of a training population.}

\item{yt}{A numeric vector of phenotypic values corresponding to individuals of allocation of a training set.}

\item{R0}{The known spatial correlation matrix for a specific field. Consider it independent for units if NULL.}

\item{nIters}{Number of iterations.}

\item{tolParInv}{A tolerance parameter of matrix inversion.}

\item{method}{2 estimation methods "REML" and "BGS" are provided.}
}
\value{
This function will return a numeric vector represents the correponding GEBVs for entire breeding population.
}
\description{
Obtain the GEBVs for a given training set by REML or Bayesian estimation.
}
