% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIA.R
\name{GIA}
\alias{GIA}
\title{Geometrically Isomorphic Allocation for rectangular field}
\usage{
GIA(v, gia.rf)
}
\arguments{
\item{v}{is a vector}

\item{gia.rf}{A integer vector of length 2 indicating nrow and ncol of a rectangular field.}
}
\description{
Deal with the geometrically isomorphic allocation by reflecting and rotating to standard order.
A 2 by 3 example is demonstrated.
The allocation matrix(c(1:6), 2, 3, byrow = TRUE) is equivalent to:
\enumerate{
\item matrix(c(3,2,1,6,5,4), 2, 3, byrow = TRUE) by reflection,
\item matrix(c(4,5,6,1,2,3), 2, 3, byrow = TRUE) by reflection,
\item matrix(c(6,5,4,3,2,1), 2, 3, byrow = TRUE) by rotation.
NULL if consider those are different.
}
}
