% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorghum.R
\docType{data}
\name{Ksg}
\alias{Ksg}
\title{Sorghum Genome Dataset}
\format{
A matrix with 451 rows and columns.
}
\source{
\url{https://doi.org/10.1007/s00122-017-3033-y}
}
\usage{
Ksg
}
\description{
The demo genome dataset originates from Fernandes et al. (2018).
This kinship matrix for 451 individuals is calculated by 56,299 SNPs.
article: Efficiency of Multi-trait, Indirect, and Trait-assisted Genomic
Selection for Improvement of Biomass Sorghum
}
\keyword{datasets}
