% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/troprice.R
\docType{data}
\name{Ktr}
\alias{Ktr}
\title{Tropical Rice Genome Dataset}
\format{
A matrix with 328 rows and columns.
}
\source{
\url{https://doi.org/10.1371/journal.pgen.1004982}
}
\usage{
Ktr
}
\description{
The demo genome dataset origined from Spindel et al. (2015).
This kinship matrixfor 328 individuals is calculated by 10,772 SNPs.
article: Genomic Selection and Association Mapping in Rice (Oryza sativa):
Effect of Trait Genetic Architecture, Training Population Composition,
Marker Number and Statistical Model on Accuracy of Rice Genomic Selection
in Elite, Tropical Rice Breeding Lines.
}
\keyword{datasets}
