% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheat.R
\docType{data}
\name{Kwh}
\alias{Kwh}
\title{Wheat Genome Dataset}
\format{
A matrix with 313 rows and columns.
}
\source{
\url{https://doi.org/10.3390/genes10090669}
}
\usage{
Kwh
}
\description{
The demo genome dataset originates from Kirstensen et al. (2019).
This kinship matrix for 313 individuals is calculated by 11,214 SNPs.
article: Genomic Prediction and Genome-Wide Association Studies of Flour
Yield and Alveograph Quality Traits Using Advanced Winter Wheat Breeding
Material.
}
\keyword{datasets}
