% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkR.R
\name{mkR}
\alias{mkR}
\title{mkR}
\usage{
mkR(gia.rf, rho = 0.5, fun = "power")
}
\arguments{
\item{gia.rf}{A integer vector of length 2 indicating nrow and ncol of a retangular field for geometrically isomorphic allocation. See function \code{GIA}.}

\item{rho}{A numeric value for spatial correlation parameter ranges from 0 to 1.}

\item{fun}{3 spatial correlation functions "power", "exponential", and "Gaussian" are provided.}
}
\description{
A function to generate spatial correlation matrix R for a given retangular field.
}
