% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDCG.R
\name{NDCG}
\alias{NDCG}
\title{Normalized Discounted Cumulative Gain}
\usage{
NDCG(v, vhat, k, gain = "linear")
}
\arguments{
\item{v}{A numeric vector representing true values for breeding population.}

\item{vhat}{A numeric vector representing predicted values for breeding
population.}

\item{k}{k An integer representing the number of individuals that will be
selected in GS procedure.}
}
\description{
An evaluation metric that measures the performance of predicted top k
individuals by giving discounted weights to their corresponding true values.
}
