% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSratio.R
\name{RSratio}
\alias{RSratio}
\title{Rank Sum Ratio}
\usage{
RSratio(v, vhat, k)
}
\arguments{
\item{v}{A numeric vector representing true values for breeding population.}

\item{vhat}{A numeric vector representing predicted values for breeding
population.}

\item{k}{An integer representing the number of individuals that will be
selected in GS procedure.}
}
\description{
An evaluation metric describing the true location of the predicted top k
individuals within breeding population.
}
