% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSOSGS.R
\name{TSOSGS}
\alias{TSOSGS}
\title{Training Set Optimization for Spatial Genomic Selection}
\usage{
TSOSGS(
  Kb,
  nt,
  test = NULL,
  R = NULL,
  lambda = NULL,
  gia.rf = NULL,
  method = "TS",
  ns = NULL,
  threshold = 10^-4,
  nIters = 10^4,
  tolParInv = 10^-4,
  mc.cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{Kb}{A nb*nb Kinship matrix of breeding population.}

\item{nt}{A integer of training set size as desired.}

\item{test}{A vector indicating target population.}

\item{R}{A nt*nt covariance matrix of random error for spatial correlation.}

\item{lambda}{A single value for computing CDmean. Defaulted to 1.}

\item{gia.rf}{A integer vector of length 2 indicating nrow and ncol of a
retangular field for geometrically isomorphic allocation. See function \code{GIA}.}

\item{method}{Two-Stage "TS" or One-Stage "OS" method are provided.}

\item{ns}{Number of solutions for one generation in GA.}

\item{threshold}{An acceptable threshold of improvement of nearest 100
iterations in GA stop criterion.}

\item{nIters}{Number of iterations that must be reached in GA.}

\item{tolParInv}{A tolerance parameter of matrix inversion.}

\item{mc.cores}{Number of used cores. See funtion \code{mclapply} in package
\code{parallel}.}

\item{verbose}{Show details in console if TRUE and vice versa.}
}
\value{
This function will return the optimal training population from
candidate with the best allocation in field.
}
\description{
This function provides Two-Stage or One-Stage Genetic Search algorithm for
Training Set Optimization for Spatial Genomic Selection under an additive
spatial GBLUP model.
GBLUP model: yt ~ mu\emph{1 + gt + e0, where
yt is phenotypic values for a given training population;
mu is fixed parameter of population mean;
1 denotes column vector with length nt;
gt is genotypic values of corresponding individuals for training set;
e0 is random error for a specific field with known spatially correlated
experimental units.
gt ~ MVN(0, sigma_g^2}Kt),
e0 ~ MVN(0, sigma_e^2*R0).
}
\examples{
# Using built-in dataset as demonstration.
Kb <- data(tropical.rice)
nt <- 50
gia.rf <- c(5, 10)
R0 <- mkR(c(5, 10), rho = 0.8)
# Optimization without spatial correlation.
TSOSGS(Kb, nt)
# Two-Stage method considering geometrically isomorphic allocation.
TSOSGS(Kb, nt, gia.rf = gia.rf)
# One-Stage method with specific spatial correlation matrix R.
TSOSGS(Kb, nt, R = R0, method = "OS")
}
